Basic separation into two modules.
